# ARM64 NEON SIMD huffman encoding