Refer to #44, #47 and #49 for more details.

Timings on the Cavium ThunderX ARM64 board:
master d38b4f21ec5baabb448cd9ffa078fa9150d54af2 (edited 2016-01-17)
```
vgl_6434_0018a.bmp 100 444 10.805636 Megapixels/sec
vgl_6434_0018a.bmp 95 444 13.092066 Megapixels/sec
vgl_6434_0018a.bmp 95 422 16.349982 Megapixels/sec
vgl_6434_0018a.bmp 95 420 20.339437 Megapixels/sec
artificial.ppm 100 444 15.478101 Megapixels/sec
artificial.ppm 95 444 20.059311 Megapixels/sec
artificial.ppm 95 422 26.932211 Megapixels/sec
artificial.ppm 95 420 33.242520 Megapixels/sec
nightshot_iso_100.ppm 100 444 12.438681 Megapixels/sec
nightshot_iso_100.ppm 95 444 19.313331 Megapixels/sec
nightshot_iso_100.ppm 95 422 25.977781 Megapixels/sec
nightshot_iso_100.ppm 95 420 32.086176 Megapixels/sec

vgl_6434_0018a.bmp 10 444 18.169759 Megapixels/sec
vgl_6434_0018a.bmp 50 444 16.888690 Megapixels/sec
vgl_6434_0018a.bmp 95 444 13.120563 Megapixels/sec
artificial.ppm 10 444 21.620149 Megapixels/sec
artificial.ppm 50 444 21.266031 Megapixels/sec
artificial.ppm 95 444 20.155321 Megapixels/sec
nightshot_iso_100.ppm 10 444 21.657411 Megapixels/sec
nightshot_iso_100.ppm 50 444 21.346780 Megapixels/sec
nightshot_iso_100.ppm 95 444 19.325315 Megapixels/sec
```

With this PR  (edited 2016-01-17):
```
vgl_6434_0018a.bmp 100 444 10.178003 Megapixels/sec
vgl_6434_0018a.bmp 95 444 12.460482 Megapixels/sec
vgl_6434_0018a.bmp 95 422 15.655713 Megapixels/sec
vgl_6434_0018a.bmp 95 420 19.546664 Megapixels/sec
artificial.ppm 100 444 13.969191 Megapixels/sec
artificial.ppm 95 444 17.256691 Megapixels/sec
artificial.ppm 95 422 23.585533 Megapixels/sec
artificial.ppm 95 420 29.322864 Megapixels/sec
nightshot_iso_100.ppm 100 444 12.287634 Megapixels/sec
nightshot_iso_100.ppm 95 444 16.835227 Megapixels/sec
nightshot_iso_100.ppm 95 422 23.063786 Megapixels/sec
nightshot_iso_100.ppm 95 420 28.802179 Megapixels/sec

vgl_6434_0018a.bmp 10 444 15.608763 Megapixels/sec
vgl_6434_0018a.bmp 50 444 14.918838 Megapixels/sec
vgl_6434_0018a.bmp 95 444 12.457544 Megapixels/sec
artificial.ppm 10 444 18.014999 Megapixels/sec
artificial.ppm 50 444 17.840014 Megapixels/sec
artificial.ppm 95 444 17.257215 Megapixels/sec
nightshot_iso_100.ppm 10 444 18.039461 Megapixels/sec
nightshot_iso_100.ppm 50 444 17.854830 Megapixels/sec
nightshot_iso_100.ppm 95 444 16.807877 Megapixels/sec
```

With this PR + ZigZag `tbl/tblx` NEON from #47 (edited 2016-01-17):
```
vgl_6434_0018a.bmp 100 444 5.193216 Megapixels/sec
vgl_6434_0018a.bmp 95 444 5.725389 Megapixels/sec
vgl_6434_0018a.bmp 95 422 7.642293 Megapixels/sec
vgl_6434_0018a.bmp 95 420 9.835368 Megapixels/sec
artificial.ppm 100 444 6.056353 Megapixels/sec
artificial.ppm 95 444 6.594384 Megapixels/sec
artificial.ppm 95 422 9.552110 Megapixels/sec
artificial.ppm 95 420 12.386251 Megapixels/sec
nightshot_iso_100.ppm 100 444 5.719931 Megapixels/sec
nightshot_iso_100.ppm 95 444 6.534959 Megapixels/sec
nightshot_iso_100.ppm 95 422 9.447844 Megapixels/sec
nightshot_iso_100.ppm 95 420 12.272391 Megapixels/sec

vgl_6434_0018a.bmp 10 444 6.309603 Megapixels/sec
vgl_6434_0018a.bmp 50 444 6.193853 Megapixels/sec
vgl_6434_0018a.bmp 95 444 5.729896 Megapixels/sec
artificial.ppm 10 444 6.711948 Megapixels/sec
artificial.ppm 50 444 6.688382 Megapixels/sec
artificial.ppm 95 444 6.600798 Megapixels/sec
nightshot_iso_100.ppm 10 444 6.715829 Megapixels/sec
nightshot_iso_100.ppm 50 444 6.695091 Megapixels/sec
nightshot_iso_100.ppm 95 444 6.480968 Megapixels/sec
```

Timings on iPhone 5S:
master d38b4f21ec5baabb448cd9ffa078fa9150d54af2 (edited 2016-01-17)  
```
artificial.ppm 100 444 33.969899 Megapixels/sec
artificial.ppm 95 444 43.914393 Megapixels/sec
artificial.ppm 95 422 65.801160 Megapixels/sec
artificial.ppm 95 420 82.042021 Megapixels/sec
nightshot_iso_100.ppm 100 444 19.755143 Megapixels/sec
nightshot_iso_100.ppm 95 444 41.288920 Megapixels/sec
nightshot_iso_100.ppm 95 422 58.889601 Megapixels/sec
nightshot_iso_100.ppm 95 420 72.865980 Megapixels/sec

artificial.ppm 10 444 56.494088 Megapixels/sec
artificial.ppm 50 444 53.840024 Megapixels/sec
artificial.ppm 95 444 43.914393 Megapixels/sec
nightshot_iso_100.ppm 10 444 58.658252 Megapixels/sec
nightshot_iso_100.ppm 50 444 51.388827 Megapixels/sec
nightshot_iso_100.ppm 95 444 41.288920 Megapixels/sec
```

With this PR (edited 2016-01-17):
```
artificial.ppm 100 444 44.980973 Megapixels/sec
artificial.ppm 95 444 56.373134 Megapixels/sec
artificial.ppm 95 422 85.829465 Megapixels/sec
artificial.ppm 95 420 106.047026 Megapixels/sec
nightshot_iso_100.ppm 100 444 32.613519 Megapixels/sec
nightshot_iso_100.ppm 95 444 53.303646 Megapixels/sec
nightshot_iso_100.ppm 95 422 78.063981 Megapixels/sec
nightshot_iso_100.ppm 95 420 97.723906 Megapixels/sec

artificial.ppm 10 444 66.909933 Megapixels/sec
artificial.ppm 50 444 64.246335 Megapixels/sec
artificial.ppm 95 444 56.373134 Megapixels/sec
nightshot_iso_100.ppm 10 444 69.870525 Megapixels/sec
nightshot_iso_100.ppm 50 444 67.061232 Megapixels/sec
nightshot_iso_100.ppm 95 444 53.303646 Megapixels/sec
```

With this PR + ZigZag `tbl/tblx` NEON from #47 (edited 2016-01-17):
```
artificial.ppm 100 444 47.066281 Megapixels/sec
artificial.ppm 95 444 59.685320 Megapixels/sec
artificial.ppm 95 422 89.740403 Megapixels/sec
artificial.ppm 95 420 111.626894 Megapixels/sec
nightshot_iso_100.ppm 100 444 33.706232 Megapixels/sec
nightshot_iso_100.ppm 95 444 55.344628 Megapixels/sec
nightshot_iso_100.ppm 95 422 82.148109 Megapixels/sec
nightshot_iso_100.ppm 95 420 103.424884 Megapixels/sec

artificial.ppm 10 444 70.874391 Megapixels/sec
artificial.ppm 50 444 67.772399 Megapixels/sec
artificial.ppm 95 444 59.685320 Megapixels/sec
nightshot_iso_100.ppm 10 444 73.730281 Megapixels/sec
nightshot_iso_100.ppm 50 444 70.327086 Megapixels/sec
nightshot_iso_100.ppm 95 444 55.344628 Megapixels/sec
```