@dcommander 2016-01-16
I can confirm that the ARM64 Huffman SIMD code, as implemented, slows performance on ThunderX by 12% on average for your first implementation and by more than 2x for the zig-zag implementation.

Please do the following:
* Rebase this branch against master, since the rest of the ARM64 NEON routines have been checked in.
* Please do not close this PR.  I prefer that we maintain one discussion thread for each feature instead of you opening a new PR each time you change a commit.  If you change the commits on this branch or add new commits, then it will update this PR, which is what I want.  Let's keep all discussion of ARM64 Huffman SIMD performance here.  When there are multiple PRs for the same feature, it becomes very difficult for me to track the information.
* Please change the title to something that indicates that this PR is solely for ARM64 Huffman encoding.