@mayeut 2016-01-17
@ssvb, seems you're using the Cavium ThunderX. Not to closer you're results I've just logged off.
Can you tell  me when I can use it ?  thanks