@mayeut 2016-01-17
> Please do not close this PR. I prefer that we maintain one discussion thread for each feature instead of you opening a new PR each time you change a commit. If you change the commits on this branch or add new commits, then it will update this PR, which is what I want. Let's keep all discussion of ARM64 Huffman SIMD performance here. When there are multiple PRs for the same feature, it becomes very difficult for me to track the information.

I know that new commits are automatically added to the PR, this how I worked on #44. However I didn't find a way to squash all commits from a PR branch (published) into 1 commit (you asked for) in the same branch. That's why I closed  & reopened PRs. Googling around didn't help me with this, maybe I didn't search right.