@ssvb 2016-01-17
@mayeut Sorry, I was just using ThunderX for some quick experiments with unaligned reads/writes. I have logged off now.

As for squashing commits, you can https://www.kernel.org/pub/software/scm/git/docs/user-manual.html#cleaning-up-history in a branch (using "git reset --hard" and other git commands) and then force push it to github.