@dcommander 2016-01-17
To combine commits, use git rebase -i, then force push to the branch.

But I was probably wrong to ask those commits to be combined earlier. I assumed they were in a more final state that I could merge without much modification, but it seems as if all of this stuff is still going to be somewhat in flux until Siarhei and I get ARM64 hardware in house.

I would suggest for now that we do the best we can with what we have available (iPhone and ThunderX), and further testing and optimization can occur later if necessary. I need to keep this moving forward. It seems like the ThunderX is sort of a lowest common denominator, so I suspect that, in most cases, if code performs well on that chip and on the iPhone, it will probably perform well on other chips.