@ssvb 2016-01-18
The huffman NEON code is a pretty impressive nice optimization.  Just one minor nitpick, unrelated to the correctness and performance. Would it be possible to use human readable aliases for registers via .req/.unreq directives just like the rest of the ARM assembly code does?

And regarding the ThunderX performance problems on huffman code. I have already mentioned this in another pull request discussion. If it turns out that there is no way to have a single universal blended implementation which works optimal on all hardware, then the runtime CPU features detection code can be extended to also identify microarchitecture variants. On ARM the information about the CPU core is pretty much standardized and available in the [MIDR](http://infocenter.arm.com/help/topic/com.arm.doc.ddi0344k/Bgbiddeb.html) system register (it has the vendor id, the core variant id, the major and minor revisions of the core). On Linux, the relevant parts of the MIDR register are exposed in /proc/cpuinfo and can be extracted from it. Here is one example of a script, which attempts to decode this information: https://github.com/dword1511/shell-scripts/blob/master/arminfo
The disadvantage is of course the addition of extra code paths, which need to be exercised during 'make test' (this can be done automatically by using the JSIMD_* environment variable and running multiple passes with different settings).