@dcommander 2016-01-18
I would be opposed to having multiple SIMD implementations for the same architecture, but I would not be opposed to detecting ThunderX at run time and falling back to the C Huffman implementation when running on that specific chip.  Primarily, this approach would ensure that all code paths get tested irrespective of the presence of ThunderX (the C Huffman encoder will still be used on non-x86 and non-ARM architectures, and of course anyone still using 1.4.x will be using that encoder.)  I unfortunately can't assume that 'make test' will cover all bases.  It never fails that someone comes up with an obscure security issue when fuzzing the code (Mozilla does a lot of that), and it always seems to affect the Huffman code-- however, only a couple of instances of this have affected the encoder (usually they catch issues by creating corrupt JPEG files and trying to decode them.)

From a maintenance point of view, it bothers me to have a code path that is rarely used in real-world situations, because it is often those real-world situations that reveal obscure bugs.  Thus, my opinion on this is:  if you can't (or don't want to) make the Huffman SIMD code run well on ThunderX, let's just use the C Huffman code on that chip.