@mayeut 2016-01-18
@ssvb agreed for human readable aliases. Since registers were used for different things throughout the code in the initial implementation, it was easier to check for mis reuse of a register. Now that it's cleaner, I can move it further by using human readable aliases.

@dcommander, my knowledge to optimize for a specific micro architecture is limited (in-order, dual-issue, ...) and even with @ssvb input, I'm still missing some bits & pieces. If a "cycle counter" exists for this SOC, like it exists for the [Cortex A8](http://pulsar.webshaker.net/ccc/result.php?lng=fr), I can try do something by relying on this tool. Otherwise, I won't be able to do optimize this for the ThunderX.

If decided not to optimize for ThunderX, I think it's best to wait for you to receive you ARM64 boards in order to choose between linear load + `tbl/tbx` zigzag (which gives the best performances on iPhone 5S) and this implementation.